+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [Unreleased]
 
 ### Added
+- Add auto-transition feature for all workflows
 
 ### Changed
+- Update styling and colours
 
 ### Fixed